Update URLs and email addresses for fork
Update project URLs, bug reporting URLs and email addresses in several
places to account for the fork instead of the original Corebird. My
intention here is not to erase Timm Bäder from the project, but to
reduce the amount of requests he gets for a project that he appears to
have no more interest in. I’ll be happy to revert any of these changes
at his request.