Remove compatibility with old config directory
Corebird used to read its settings from ~/.corebird before being changed
to use the XDG config directory by default; for compatibity, it still
supported the ~/.corebird path. In Nextbird, we can drop this,
simplifying configuration handling. (This was briefly discussed in #5,
with the result that it’s not very important for Nextbird to
automatically import Corebird’s configuration.)